fix(resource): 🐛 filter ecPois by osmfeatures_id and convert to array
Added a filter to the `getRelatedPois` method to exclude `ecPois` with a non-null `osmfeatures_id`. This ensures that POIs obtained from OSM are not added to the related list. Additionally, the result is converted to an array to maintain consistency.

refactor(job): ♻️ simplify ecTrackModelClass configuration

Removed the default value for `ecTrackModelClass` in `UpdateEcTrackAwsJob` to rely solely on the configuration file. This change ensures that the model class used is always the one defined in the configuration, promoting consistency across different environments.