feat(app): ✨ add options for download tracks and tiles
Added new Boolean fields to enable or disable the display of download options for tracks in GPX, KML, GEOJSON formats, and tiles on the map. These options can be configured in the Nova App resource.

Additionally, added text fields for Play Store and App Store links, which are hidden from the index but provide helpful links to the app on respective stores.

Updated the AppConfigService to include the new 'showDownloadTilesButton' option, ensuring that it is set correctly based on the app's properties.