feat(app): ✨ add options for download tracks and tiles oc:6581 (#147)
* feat(app): ✨ add options for download tracks and tiles

Added new Boolean fields to enable or disable the display of download options for tracks in GPX, KML, GEOJSON formats, and tiles on the map. These options can be configured in the Nova App resource.

Additionally, added text fields for Play Store and App Store links, which are hidden from the index but provide helpful links to the app on respective stores.

Updated the AppConfigService to include the new 'showDownloadTilesButton' option, ensuring that it is set correctly based on the app's properties.

* refactor(App): 🔄 rename property for download tiles option

Renamed 'show_download_tiles_button' to 'show_download_tiles' in the App class to maintain consistency and clarity across the codebase. Updated the associated logic in AppConfigService to reflect this change.

Additionally, added a TODO comment to indicate that 'showDownloadTilesButton' is slated for removal in the next release, as it is only used by version 3.1.6.