fix: update Name field logic oc: 6625 (#154)
* fix: update Name field logic oc: 6625

Modifies the Name field to provide fallback logic for missing
values in properties, ensuring more robust data display.

Relates to oc_6625

* refactor: simplify name field access oc: 6625

Replaces closure with a direct field access method for the 'Name'
field, improving clarity and performance.

Relates to oc_6625

* fix: update name field logic oc: 6625

Modifies the logic for the 'Name' field to use a fallback
from 'form title' if 'name' is unavailable. Also, makes
the field read-only for additional data protection.

Relates to oc_6625

* Fix styling

* fix: update Name field logic oc: 6625

Modifies the Name field to provide fallback logic for missing
values in properties, ensuring more robust data display.

Relates to oc_6625

* refactor: simplify name field access oc: 6625

Replaces closure with a direct field access method for the 'Name'
field, improving clarity and performance.

Relates to oc_6625

* fix: update name field logic oc: 6625

Modifies the logic for the 'Name' field to use a fallback
from 'form title' if 'name' is unavailable. Also, makes
the field read-only for additional data protection.

Relates to oc_6625

* refactor(Nova): ♻️ simplify name retrieval logic in AbstractUgcResource

- Replace the ternary operator with the `data_get` helper function for cleaner and more readable code.
- The `data_get` function is used to safely retrieve nested data, reducing potential errors when accessing properties.

---------

Co-authored-by: peppedeka <peppedeka@users.noreply.github.com>
Co-authored-by: bongiu <peppedeka@gmail.com>